Move metadata store classes to the right package (service -> store).
